refactor(factories): create reusable shorthand factory
